Clearer error on Py<3.6.
